requirements.txt
glfw==2.10.0
numpy==2.3.4
PyOpenGL==3.1.10
